+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Various false positive attribute drifts ([#344](https://github.com/Scalr/terraform-provider-scalr/pull/344))
 
+### Changed
+
+- `scalr_variable`: added deprecation warning when `hcl` attribute is set to `true` for shell variable
+
 ## [2.1.0] - 2024-09-06
 
 ### Added